% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_data.R
\docType{data}
\name{binary_data_for_example}
\alias{binary_data_for_example}
\title{Pedagogic data set with binary ratings}
\format{
An initial data frame of dimensions 8 x 20 containing the binary ratings on 8 stimuli provided by a panel of 20 raters. The initial data frame is supplemented by 3 covariates providing information about the raters (Gender, Age, Freq.use; the first 2 categorical and the last one continuous); and by 5 covariates providing information about the stimuli (Citrus.fruits, Vanilla, Wood, Lotus, Packaging; the first 4 continuous and the last one categorical).
}
\usage{
binary_data_for_example
}
\description{
This data set is a small artificial pedagogic data set.
}
\examples{
data("continuous_data_for_example")

}
\author{
Margot Brard \email{margot.brard@yahoo.fr}
}
\keyword{datasets}
