% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{charact_cluster_rater}
\alias{charact_cluster_rater}
\title{Function charact_cluster_rater}
\usage{
charact_cluster_rater(
  i,
  mat.partition,
  res.pca,
  id_info_rater,
  dta.sauv,
  mat.resids,
  type_info_rater,
  id_info_stim
)
}
\description{
Function charact_cluster_rater
}
