% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{charact_cluster_stim}
\alias{charact_cluster_stim}
\title{Function charact_cluster_stim}
\usage{
charact_cluster_stim(
  i,
  mat.partition,
  id_info_rater,
  id_info_stim,
  dta.sauv,
  dta,
  type_info_stim,
  melted.data.clusters
)
}
\description{
Function charact_cluster_stim
}
