% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{compute_dist_null}
\alias{compute_dist_null}
\title{Function compute_dist_null}
\usage{
compute_dist_null(j, list.data.null, K, model, model2)
}
\arguments{
\item{j}{j}

\item{list.data.null}{list.data.null}

\item{K}{K}

\item{model}{model}

\item{model2}{model2}
}
\description{
Function compute_dist_null
}
