% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_agreeclust_binary.R
\name{get_agreeclust_bin}
\alias{get_agreeclust_bin}
\title{Agreement-based clustering of binary ratings}
\usage{
get_agreeclust_bin(
  dta,
  model = "rating ~ rater + stimulus",
  max_clust = 10,
  approx_null = TRUE,
  paral_null = TRUE,
  consol = TRUE,
  id_info_rater = NULL,
  type_info_rater = NULL,
  id_info_stim = NULL,
  type_info_stim = NULL,
  graph = TRUE,
  ext_dev_Rstudio = FALSE
)
}
\arguments{
\item{dta}{A binary matrix or data frame of dimensions S x R (S=number of stimuli, R=number of raters).}

\item{model}{A formula for the model. Either 'rating ~ rater + stimulus' (default), 'rating ~ rater', or 'rating ~ 1'.}

\item{max_clust}{An integer specifying the maximum number of clusters of raters. By default, this number is fixed to 10.}

\item{approx_null}{A boolean indicating if the null LRT distribution should be approximated using Satterthwaite's approximation. By default, the null LRT distribution is approximated.}

\item{paral_null}{A boolean indicating if the computation of the null LRT distribution should be parallelized. By default, the computation of the null LRT distribution is parallelized on nb.cores-1 cores. During the process, a text file 'TestDendrogram_processing.txt' is created.}

\item{consol}{A boolean indicating if a k-means consolidation of the partition of raters should be performed. By default, the partition is consolidated.}

\item{id_info_rater}{A vector of integer elements composed of the identification of the lines containing the supplementary information (i.e. covariates) about the raters. This argument is optional and, by default, it is fixed to NULL, meaning that dta does not contain supplementary information about the raters.}

\item{type_info_rater}{A vector of character elements composed of the type of the covariates about the raters. This vector must be of the same length that id.info.rater. A continuous covariate is associated to 'cont' and a categorical covariate is associated to 'cat'. This argument is optional and, by default, it is fixed to NULL, meaning that dta does not contain supplementary information about the raters.}

\item{id_info_stim}{A vector of integer elements composed of the identification of the columns containing the supplementary information (i.e. covariates) about the stimuli. This argument is optional and, by default, it is fixed to NULL, meaning that dta does not contain supplementary information about the stimuli.}

\item{type_info_stim}{A vector of character elements composed of the type of the covariates about the stimuli. This vector must be of the same length that id.info.stim. A continuous covariate is associated to 'cont' and a categorical covariate is associated to 'cat'. This argument is optional and, by default, it is fixed to NULL, meaning that dta does not contain supplementary information about the stimuli.}

\item{graph}{A boolean specifying if the graphical outputs should be plotted or not. By default, they are plotted.}

\item{ext_dev_Rstudio}{A boolean specifying if the graphical outputs should be plotted in the Rstudio plot pane or not.}
}
\value{
list
\itemize{
\item{profiles_residuals}{A matrix of dimensions S x R (S=number of stimuli, R=number of raters) containing the residuals profiles of the raters obtained through the modelling of the set of binary ratings.}
\item{mat_disag}{A matrix of dimensions R x R (R=number of raters) corresponding to the dissimilarity matrix between the raters.}
\item{pval_dendro}{A vector containing the probabilities associated to the statistical test realized at each level of the dendrogram.}
\item{nb_clust_found}{An integer corresponding to the number of clusters found among the panel.}
\item{partition}{A vector representing the partition of the raters (consolidated partition if consol = TRUE).}
\item{res_plot_segment}{All the graphical results of the segmentation.}
\item{res_pca}{All the results of the PCA.}
\item{charact_clust}{The results of the description of the clusters by information describing the raters and/or the stimuli.}
}
}
\description{
Agreement-based clustering of binary ratings
}
\examples{
data(binary_data_for_example)
res_pedag <- get_agreeclust_bin(dta = binary_data_for_example,
                                id_info_rater = 9 : nrow(binary_data_for_example),
                                type_info_rater = c(rep("cat", 2), "cont"),
                                id_info_stim = 21 : ncol(binary_data_for_example),
                                type_info_stim = c(rep("cont", 4), "cat"),
                                paral_null = FALSE
                                )
res_pedag

\dontrun{
data(goodgesture)
res_goodgesture <- get_agreeclust_bin(dta = goodgesture,
                                      model = "rating ~ rater + stimulus",
                                      id_info_rater = 40 : nrow(goodgesture),
                                      type_info_rater = rep("cat", 4),
                                      id_info_stim = 73 : ncol(goodgesture),
                                      type_info_stim = c(rep("cat", 3), rep("cont", 11)),
                                      paral_null = FALSE
                                      )
}
}
