% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{GlmSum}
\alias{GlmSum}
\title{Function remove_outliers}
\usage{
GlmSum(formula, data, na.action = na.omit)
}
\arguments{
\item{formula}{formula}

\item{data}{data}

\item{na.action}{na.action}
}
\description{
Function remove_outliers
}
