\name{pedagdataCont}
\alias{pedagdataCont}
\docType{data}
\title{
  Pedagogic data set with continuous ratings
}
\description{
This data set is a small artificial pedagogic data set.
}
\usage{data("pedagdataCont")}
\format{
An initial data frame of dimensions 8 x 20 containing the continuous ratings on 8 stimuli provided by a panel of 20 raters. The initial data frame is supplemented by 3 covariates providing information about the raters (Gender, Age, Freq.use; the first 2 categorical and the last one continuous); and by 5 covariates providing information about the stimuli (Citrus.fruits, Vanilla, Wood, Lotus, Packaging; the first 4 continuous and the last one categorical).
}
\author{Margot Brard \email{margot.brard@agrocampus-ouest.fr}}

\examples{
data(pedagdataCont)
}
\keyword{datasets}
