% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_agreeclust.R
\name{plot_agreeclust}
\alias{plot_agreeclust}
\title{Graphical representation of the structure of disagreement among the panel of raters}
\usage{
plot_agreeclust(res, choice = "all", interact = FALSE, col_clust = NULL, axis = c(1, 2), name_rater = "rater", ext_dev_Rstudio = FALSE)
}
\arguments{
\item{res}{An object of class agreeclust returned by the functions of the package.}

\item{choice}{A character element specifying the graphs to be plotted ('seg' for the representation of the segmentation process (dendrogram + partitioning if a consolidation process has been implemented), 'mul' for the multidimensional representation of the structure of disagreement (PCA), 'all' for both representations. By default, both representations are plotted.)}

\item{interact}{A boolean specifying if the graphical outputs should be interactive (with plotly) or not.}

\item{col_clust}{A vector with as many color as clusters of raters. Colors can be specified by a character such as "blue" or by a hexadecimal code such as "#0D10D5". All hexadecimal codes can be found at \url{http://htmlcolorcodes.com}. By default, no colors are specified and default colors are used.}

\item{axis}{A length 2 numeric vector specifying the PCA components to be plotted. By default, the first 2 components are plotted.}

\item{name_rater}{A character element indicating how the raters should be named (e.g. "rater", "participant", "psychologists")}

\item{ext_dev_Rstudio}{A boolean specifying if the graphical outputs should be plotted in the Rstudio plot pane or not.}
}
\value{
Returns the graphical representations
}
\description{
Draws the graphs representing the structure of disagreement among the set of ratings. Tunning parameters such as the color of the clusters or the axis of PCA to be plotted can be specified.
}
\examples{
data(binary_data_for_example)
res_pedag <- get_agreeclust_bin(dta = binary_data_for_example,
                                id_info_rater = 9 : nrow(binary_data_for_example),
                                type_info_rater = c(rep("cat", 2), "cont"),
                                id_info_stim = 21 : ncol(binary_data_for_example),
                                type_info_stim = c(rep("cont", 4), "cat"),
                                paral_null = FALSE
                                )
plot_agreeclust(res_pedag, col_clust = c("blue", "green"))
plot_agreeclust(res_pedag, choice = "mul", interact = TRUE, col_clust = c("blue", "green"))
}
