% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{print.agreeclust_binary}
\alias{print.agreeclust_binary}
\title{Function print.get_agreeclust_binary}
\usage{
\method{print}{agreeclust_binary}(res)
}
\arguments{
\item{res}{res}
}
\description{
Function print.get_agreeclust_binary
}
