% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{remove_outliers}
\alias{remove_outliers}
\title{Function remove_outliers}
\usage{
remove_outliers(x, na.rm = TRUE, ...)
}
\arguments{
\item{x}{x}

\item{na.rm}{na.rm}

\item{...}{...}
}
\description{
Function remove_outliers
}
