% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{compute_pval}
\alias{compute_pval}
\title{Function compute_pval}
\usage{
compute_pval(
  K,
  dendrogram,
  melted.data,
  model,
  model2,
  approx_null,
  dta,
  paral_null
)
}
\arguments{
\item{K}{K}

\item{dendrogram}{dendrogram}

\item{melted.data}{melted.data}

\item{model}{model}

\item{model2}{model2}

\item{approx_null}{approx_null}

\item{dta}{dta}

\item{paral_null}{paral_null}
}
\description{
Function compute_pval
}
