\name{pedagdataBin}
\alias{pedagdataBin}
\docType{data}
\title{
  Pedagogic data set with binary ratings
}
\description{
This data set is a small artificial pedagogic data set.
}
\usage{data("pedagdata")}
\format{
An initial data frame of dimensions 8 x 20 containing the binary ratings on 8 stimuli provided by a panel of 20 raters. The initial data frame is supplemented by 3 covariates providing information about the raters (Gender, Age, Freq.use; all categorical); and by 5 covariates providing information about the stimuli (Citrus.fruits, Vanilla, Wood, Lotus, Packaging; the first 4 continuous and the last one categorical).
}
\examples{
data(pedagdataBin)
}
\keyword{datasets}
